+++ src/platform/macos/util/mod.rs
@@ -1,3 +1,7 @@
+mod cursor;
+
+pub use self::cursor::*;
+
 use cocoa::appkit::NSWindowStyleMask;
 use cocoa::base::{id, nil};
 use cocoa::foundation::{NSRect, NSUInteger};